Split account fetch process
Change how the details are published. Previously a single lambda would
gather the information and push to Slack. This change now has a lambda
which publishes a data structure (JSON) to AWS SNS, which is subscribed
by another lambda, which transform the message and posts to Slack.

The purpose of this change is to allow the data to fan-out easily to
multiple consumers, by way of subscribing to the SNS topic