Add limit to Slack message
Add a limit of the top 50 accounts to be sent to the Slack message.
Why? Because Slack has started posting 2 messages for the 60ish accounts we have. The actually limit seems to be 54 lines, but unsure if this is lines or a character limit, and 50 is a nice round number