Fix lambda return type
Per reference docs here, https://docs.aws.amazon.com/lambda/latest/dg/python-handler.html; "If the handler returns objects that can't be serialized by json.dumps, the runtime returns an error."
This was causing the lamdba to view the execution as a failure, and thus be subjected to the standard 3x retry. But the process was all working correctly, so this resulted in the reply message being posted 3x into Slack.
Change here is to return the HTTP response data instead, not that anything is actually needed, as the lambda is called asynchronously via an SNS trigger